some code cleanup

+++ airootfs/root/customize_airootfs.sh
@@ -93,9 +93,6 @@ if [ $RET -ne 0 ];then
     cd ..
 fi
 
-# install Oracle JRE (because JOdin will not run with OpenJDK)
-#yaourt -Q jre || su -c - $LOGINUSR "yaourt -S --noconfirm jre"
-
 # install yad
 echo -e "\nyad:"
 yaourt -Q yad || su -c - $LOGINUSR "yaourt -S --noconfirm yad"
@@ -105,7 +102,7 @@ cp /usr/share/applications/pamac-manager.desktop /home/$LOGINUSR/Desktop/
 sed -i 's/Icon=.*/Icon=gnome-software/g' /home/$LOGINUSR/Desktop/pamac-manager.desktop
 
 # disable tray to avoid bothering users for updating
-rm /etc/xdg/autostart/pamac-tray.desktop
+[ -f /etc/xdg/autostart/pamac-tray.desktop ] && rm /etc/xdg/autostart/pamac-tray.desktop
 
 # minimal web browser
 yaourt -Q otter-browser || su -c - $LOGINUSR "yaourt -S --noconfirm otter-browser"
